requirements.txt
# Core Flask dependencies (Python 3.13 compatible)
Flask>=2.3.0
Flask-CORS>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional dependencies (install only if needed)
# pandas>=2.0.0
# numpy>=1.24.0
# scikit-learn>=1.3.0
# googletrans>=4.0.0
# gtts>=2.4.0
# SpeechRecognition>=3.10.0
# geopy>=2.4.0
# ollama>=0.1.7